_documentation-blueprint/requirements.txt
# Documentation Blueprint Dependencies

# Core dependencies
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.0
pytest-cov>=4.0
